+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 0.1.7 Dev
+- fixed problem with spcecific series (9-1-1 for example)
+- fixed issue with error on scheduled M3U download
+
 ## 0.1.6 Dev
 - Moved 'rebuild M3u' menu to to 'channels' in navigation menu
 - downloading movies would mention series, fixed